Split lib into core and extras to allow separate dependency scopes
