+++ espresso-support/core/build.gradle
@@ -1,0 +1,38 @@
+buildscript {
+    repositories {
+        jcenter()
+    }
+    dependencies {
+        classpath libraries.build.androidGradle
+        classpath libraries.build.bintrayRelease
+    }
+}
+
+apply plugin: 'com.android.library'
+apply plugin: 'bintray-release'
+
+android {
+    compileSdkVersion versions.androidSdk.compile
+    buildToolsVersion versions.androidSdk.buildTools
+
+    defaultConfig {
+        minSdkVersion versions.androidSdk.min
+    }
+}
+
+dependencies {
+    provided project(':extras')
+    compile libraries.androidTestRunner
+    compile libraries.androidEspressoCore
+}
+
+publish {
+    userOrg = 'novoda'
+    repoName = 'maven'
+    groupId = 'com.novoda'
+    artifactId = 'espresso-support-core'
+    version = project.version
+    uploadName = 'espresso-support'
+    description = 'Support classes and functions for Espresso testing'
+    website = 'https://github.com/novoda/spikes/tree/master/espresso-support'
+}
