+++ espresso-support/demo/src/androidTest/java/com/novoda/movies/MovieItemViewTest.java
@@ -3,7 +3,7 @@ package com.novoda.movies;
 import android.support.test.filters.LargeTest;
 import android.support.test.runner.AndroidJUnit4;
 
-import com.novoda.espresso.ViewActivityRule;
+import com.novoda.espresso.core.ViewActivityRule;
 
 import org.junit.After;
 import org.junit.Before;
@@ -38,7 +38,6 @@ public class MovieItemViewTest {
 
     @Mock
     MovieItemView.Listener movieItemListener;
-
 
     @Before
     public void setUp() {
