+++ espresso-support/extras/src/main/AndroidManifest.xml
