+++ espresso-support/extras/src/main/java/com/novoda/espresso/ViewActivity.java
